% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_options.R
\docType{data}
\name{format_options}
\alias{format_options}
\title{Package Options}
\format{
An object of class \code{list} of length 3.
}
\usage{
format_options
}
\value{
A list of functions.
}
\description{
A list of functions for getting, setting, and appending options for
formatting.  The available options are \code{"lowercase_words"},
\code{"punctuation"}, and \code{"protected_words"}.  The default values can
be retrieved by \code{formatBibtex_defaults$get()}.
}
\keyword{datasets}
